requirements.txt
strands-ai-functions
strands-agents
pydantic>=2.0
